now 64bit integers in application code

+++ src/application/collect_and_forward/tc-based/app.c
@@ -22,7 +22,7 @@ void collect_run(const char* sensor, const char* file, unsigned dt)
     void* logw_handle = logw_wire(file);
 	void* timer_handle = timer_wire();
 
-    uint64_t now = timer_getNow(timer_handle);
+    uint32_t now = timer_getNow(timer_handle);
     while (true) {
         timer_sleep(timer_handle, now + dt);
         now += dt;
@@ -75,7 +75,7 @@ void send_run(const char* channel, const char* file1, const char* file2, unsigne
     void* logr2_handle = logr_wire(file2);
     void* send_handle = send_wire(channel);
 	void* timer_handle = timer_wire();
-    uint64_t now = timer_getNow(timer_wire);
+    uint32_t now = timer_getNow(timer_wire);
 
     while (true) {
         timer_sleep(timer_handle, now + dt);
