+++ src/operating_system/tinyos/threads/Timer.h
@@ -14,12 +14,12 @@ void* timer_wire();
  * sleep until a future point in time.
  *
  */
-error_t timer_sleep(void* handle, uint64_t until);
+error_t timer_sleep(void* handle, uint32_t until);
 
 /**
 * Return the current time.
 * @return Current time.
 */
-uint64_t timer_getNow(void* handle);
+uint32_t timer_getNow(void* handle);
 
 #endif
